Move payment method name translation functions to own file
