+++ client/my-sites/checkout/composite-checkout/types/checkout-payment-method-slug.ts
@@ -1,9 +1,4 @@
 /**
- * External dependencies
- */
-import { snakeCase } from 'lodash';
-
-/**
  * Payment method slugs as expected by composite-checkout.
  * If the composite-checkout package used typescript, this
  * would belong there.
@@ -28,20 +23,3 @@ export type CheckoutPaymentMethodSlug =
 	| 'full-credits'
 	| 'stripe-three-d-secure'
 	| 'wechat';
-
-export function translateCheckoutPaymentMethodToTracksPaymentMethod(
-	paymentMethod: CheckoutPaymentMethodSlug
-): string {
-	let paymentMethodSlug: string = paymentMethod;
-	// existing cards have unique paymentMethodIds
-	if ( paymentMethod.startsWith( 'existingCard' ) ) {
-		paymentMethodSlug = 'credit_card';
-	}
-	switch ( paymentMethodSlug ) {
-		case 'card':
-			return 'credit_card';
-		case 'apple-pay':
-			return 'web_payment';
-	}
-	return snakeCase( paymentMethodSlug );
-}
